Statically type layer
A macro helps us make an actual type definition for the Conjure layer.
This allows us to remove the boxing and work around the compiler bugs
preventing direct storage of the base layer in the client.

It does mean that you need to update two places whenever layers are
changed, but I don't think that's too bad of an issue.